Small Fixes & Changes
So I think the only blocker was the state manager script on camera.
Removing that made it work.

Beyond that, I’ve stretched the right and left panel to vertically fill
the screen, changed the camera background from skybox to solid color,
and changed the logo.  Also added a splash screen.